refactor: split CheckIn into focused sub-components (Phase 4)
CheckIn was a 581-line page mixing token validation + login + MFA
gate + shift matching with grace-window filtering + check-in mutation
+ coordinator notification fan-out + 9 different per-state UI blocks.
This PR keeps the page as a thin state-machine orchestrator (186
lines) and moves the rest into 7 focused files. Behavior is unchanged
— the existing 103 vitest tests all pass.

State-machine-driven pattern: page owns the `Step` union + supporting
fields, network actions live in checkin-actions.ts, per-state UI is
in component files, login form owns its own state.

The existing `Step` union is preserved verbatim — no reshape into a
discriminated-union state type. That cleaner shape is a Phase-5+
concern; this is a behavior-preservation refactor.

New files:
  src/lib/shift-time.ts  (+25)               isWithinPostShiftGrace
                                              helper added (peer to
                                              isCheckInOpen, NOT a
                                              replacement — different
                                              windows for different
                                              problems)
  src/lib/checkin-actions.ts                  221  Pure-ish service
                                                   functions:
                                                   validateCheckinToken,
                                                   resolveLoginIdentifier
                                                   ToEmail,
                                                   fetchTodaysMatchedShifts,
                                                   recordCheckin,
                                                   notifyCoordinatorsOfCheckin
  src/components/checkin/CheckinHeader.tsx     17  Shared logo/title
  src/components/checkin/CheckinStateScreens   98  6 small per-state
                                                   cards (Validating,
                                                   InvalidToken,
                                                   Matching, NoShift,
                                                   AlreadyCheckedIn,
                                                   CheckinProgress)
  src/components/checkin/LoginForm.tsx        156  Owns form state +
                                                   Turnstile token +
                                                   signInWithPassword +
                                                   MFA gate
  src/components/checkin/ConfirmCheckinScreen  68  Shift picker +
                                                   "Check In to All"
  src/components/checkin/SuccessScreen.tsx     53  Success card with
                                                   shift detail or
                                                   multi-slot summary

Page: 581 → 186 (-68%)
Total: 581 → 882 (+301 net; ~400 lines of logic moved out).

Sensitive-ops sequence preserved as page-level orchestration:
  1. validateCheckinToken (RPC) — gate at mount
  2. session check — if absent, transition to login
  3. login (LoginForm) — signInWithPassword + MFA gate
     (MFA gate stays inside LoginForm — second half of auth tx)
  4. fetchTodaysMatchedShifts — applies post-shift 30-min grace
  5. recordCheckin — supabase update first
  6. notifyCoordinatorsOfCheckin — best-effort fan-out

Each numbered step is a separate page-level call. No wrapper
combines steps. Validation order intact.

Time-window logic decision:
  - CheckIn's post-shift 30-min grace and shift-time.ts's pre-shift
    isCheckInOpen are bit-different windows for different problems.
    Adding isWithinPostShiftGrace as a peer in shift-time.ts (not
    a replacement) — co-location, not unification.

Type strategy:
  - useState<any>(null) for user → User | null (structural cleanup,
    same precedent as validateAddUser reshape in PR #130).
  - 2 new boundary casts in new code (both in checkin-actions.ts):
    * `((data as any[]) || []) as BookingRow[]` — embedded select
    * `notifications as never` — Json type narrowness on inserts
      (same pattern as booking-actions.ts in PR #125)

Verification: tsc --build clean, eslint --max-warnings=0 clean,
103/103 vitest passes.

============================================================
PHASE 4 COMPLETE
============================================================

Five-PR cumulative metrics:

| PR  | Page                     | Before | After | New files |
|-----|--------------------------|--------|-------|-----------|
| #125| VolunteerDashboard       |   941  |  321  |   13      |
| #127| Settings                 |  1016  |   99  |   11      |
| #129| ManageShifts             |   662  |  110  |    5      |
| #130| AdminUsers               |   678  |  260  |    8      |
| #131| CheckIn (this)           |   581  |  186  |    7      |
| TOT | Five god components      |  3878  |  976  |   44      |

Reduction: 3878 → 976 lines (-2902 lines, -75%)
New boundary casts across all five PRs: 16
  (8 in #125, 4 in #127, 1 in #129, 1 in #130, 2 in #131)

Five god components decomposed.
